+++ src/modules/orders/orders.service.ts
@@ -2,7 +2,9 @@ import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ORDER_PAGE_LENGTH } from 'common/constants/numbers';
 import { Order } from 'common/database/entities/order.entity';
+import { User } from 'common/database/entities/user.entity';
 import { LuggageTypes, OrderStatuses } from 'common/enums/enums';
+import { AssignedOrdersResponse } from 'common/types/assignedOrdersResponse';
 import { FindManyOptions, IsNull, Like, Between, Not, Repository } from 'typeorm';
 
 import { OrderServiceParams } from './types';
@@ -12,6 +14,8 @@ export class OrdersService {
   constructor(
     @InjectRepository(Order)
     private readonly orderRepository: Repository<Order>,
+    @InjectRepository(User)
+    private readonly userRepo: Repository<User>,
   ) {}
 
   async findAll({
@@ -162,6 +166,61 @@ export class OrdersService {
       }, {});
     } catch (error: unknown) {
       throw new InternalServerErrorException(error);
+    }
+  }
+
+  private assignOrdersToDrivers(drivers: User[], orders: Order[]): AssignedOrdersResponse {
+    orders.sort((a, b) => new Date(a.collection_time_start).getTime() - new Date(b.collection_time_start).getTime());
+
+    const assignments: { driver: User; orders: Order[] }[] = drivers.map((driver) => ({
+      driver,
+      orders: [],
+    }));
+
+    const notAssignedOrders: Order[] = [];
+    const driverCount = drivers.length;
+    for (const order of orders) {
+      let assigned = false;
+
+      for (let i = 0; i < driverCount; i += 1) {
+        const assignment = assignments[i];
+        const driverOrders = assignment.orders;
+
+        const hasSameStartTime = driverOrders.some(
+          (o) => new Date(o.collection_time_start).getTime() === new Date(order.collection_time_start).getTime(),
+        );
+
+        if (!hasSameStartTime) {
+          assignment.orders.push(order);
+          assigned = true;
+          break;
+        }
+      }
+
+      if (!assigned) {
+        notAssignedOrders.push(order);
+      }
+    }
+
+    return {
+      value: assignments,
+      notAssignedOrders,
+    };
+  }
+
+  async getAndAssign(driversIds: number[], ordersIds: number[]): Promise<AssignedOrdersResponse> {
+    try {
+      const getOdersQuery = this.orderRepository
+        .createQueryBuilder('order')
+        .where('order.id IN (:...ids)', { ids: ordersIds })
+        .orderBy('order.collection_date', 'ASC');
+      const getDriversQuery = this.userRepo.createQueryBuilder('user').where('user.id IN (:...ids)', { ids: driversIds });
+      const orders = await getOdersQuery.getMany();
+      const drivers = await getDriversQuery.getMany();
+
+      return this.assignOrdersToDrivers(drivers, orders);
+    } catch (error) {
+      throw new InternalServerErrorException('something went wrong');
     }
   }
 }
